smbios: corrected folder structure
